feat(kunjungan): implement dynamic unit, dokter, dan metode pembayaran
* feat: enhance asuransi pasien management with CRUD routes and validation
* feat: add error handling for delete action and improve delete confirmation process
* feat: enhance asuransi pasien management with CRUD routes and validation
* feat: add error handling for delete action and improve delete confirmation process
* feat: add unit management
* feat: implement dokter selection by unit
* feat: Old values dokter selection
* feat: enhance validation rules and update form handling for kunjungan
* feat: update asuransi validation and handling based on payment method selection
* feat: update kunjungan model and views for improved data display and validation

+++ app/Controllers/Kunjungan.php
@@ -2,10 +2,13 @@
 
 namespace App\Controllers;
 
-use App\Models\KunjunganModel;
+use App\Models\UnitModel;
 use App\Models\PasienModel;
 use App\Models\AsuransiModel;
+use App\Models\EmpOnUnitModel;
+use App\Models\KunjunganModel;
 use App\Models\AsuransiPasienModel;
+use CodeIgniter\Exceptions\PageNotFoundException;
 
 class Kunjungan extends BaseController
 {
@@ -13,7 +16,53 @@ class Kunjungan extends BaseController
     protected $pasienModel;
     protected $asuransiModel;
     protected $asuransiPasienModel;
+    protected $unitModel;
+    protected $empOnUnitModel;
     protected $validation;
+    protected $rules = [
+        'pasien_id' => [
+            'rules' => 'required|integer',
+            'errors' => [
+                'required' => 'Pasien wajib dipilih.',
+                'integer'  => 'ID pasien tidak valid.',
+            ]
+        ],
+        'unit_id' => [
+            'rules' => 'required|integer',
+            'errors' => [
+                'required' => 'Unit wajib dipilih.',
+                'integer'  => 'ID unit tidak valid.',
+            ]
+        ],
+        'emp_on_unit_id' => [
+            'rules' => 'required|integer',
+            'errors' => [
+                'required' => 'Dokter wajib dipilih.',
+                'integer'  => 'ID dokter tidak valid.',
+            ]
+        ],
+        'tanggal_kunjungan' => [
+            'rules' => 'required|valid_date',
+            'errors' => [
+                'required' => 'Tanggal kunjungan wajib diisi.',
+                'valid_date'  => 'Tanggal kunjungan tidak valid.',
+            ]
+        ],
+        'metode_pembayaran' => [
+            'rules' => 'required|string',
+            'errors' => [
+                'required' => 'Metode pembayaran wajib diisi.',
+                'string'  => 'Metode pembayaran tidak valid.',
+            ]
+        ],
+        'keluhan' => [
+            'rules' => 'required|string|max_length[255]',
+            'errors' => [
+                'required' => 'Keluhan wajib diisi.',
+                'string'  => 'Keluhan tidak valid.',
+            ]
+        ],
+    ];
 
     public function __construct()
     {
@@ -21,6 +70,8 @@ class Kunjungan extends BaseController
         $this->pasienModel = new PasienModel();
         $this->asuransiModel = new AsuransiModel();
         $this->asuransiPasienModel = new AsuransiPasienModel();
+        $this->unitModel = new UnitModel();
+        $this->empOnUnitModel = new EmpOnUnitModel();
         $this->validation = \Config\Services::validation();
     }
 
@@ -44,12 +95,23 @@ class Kunjungan extends BaseController
         ]);
     }
 
+    public function dokterByUnit($unitId)
+    {
+        $empOnUnit = $this->empOnUnitModel->getWithEmployee()
+            ->where('unit_id', $unitId)
+            ->findAll();
+
+        return $this->response->setJSON($empOnUnit);
+    }
+
 
     public function create()
     {
 
         $pasien = $this->pasienModel->select()->findAll();
         $asuransi = $this->asuransiModel->select()->findAll();
+        $unit = $this->unitModel->select()->findAll();
+        $empOnUnit = $this->empOnUnitModel->getWithEmployee()->findAll();
 
 
         $asuransiPasien = $this->asuransiPasienModel->select('asuransi_pasien.*, pasien.nik as nik, pasien.nama as nama_pasien, asuransi.nama_asuransi')
@@ -59,69 +121,46 @@ class Kunjungan extends BaseController
 
         if ($this->request->getMethod() === 'POST') {
 
-            $rules = [
-                'pasien_id' => [
-                    'rules' => 'required|integer',
-                    'errors' => [
-                        'required' => 'Pasien wajib dipilih.',
-                        'integer'  => 'ID pasien tidak valid.',
-                    ]
-                ],
-                'asuransi_pasien_id' => [
+            $rules = $this->rules;
+
+            $metode = $this->request->getPost('metode_pembayaran');
+
+            if ($metode === 'asuransi') {
+                $rules['asuransi_pasien_id'] = [
                     'rules' => 'required|integer',
                     'errors' => [
                         'required' => 'Asuransi wajib dipilih.',
                         'integer'  => 'ID asuransi tidak valid.',
                     ]
-                ],
-
-                'tanggal_kunjungan' => [
-                    'rules' => 'required|valid_date',
-                    'errors' => [
-                        'required' => 'Tanggal kunjungan wajib diisi.',
-                        'valid_date'  => 'Tanggal kunjungan tidak valid.',
-                    ]
-                ],
-                'metode_pembayaran' => [
-                    'rules' => 'required|string',
-                    'errors' => [
-                        'required' => 'Metode pembayaran wajib diisi.',
-                        'string'  => 'Metode pembayaran tidak valid.',
-                    ]
-                ],
-                'keluhan' => [
-                    'rules' => 'required|string|max_length[255]',
-                    'errors' => [
-                        'required' => 'Keluhan wajib diisi.',
-                        'string'  => 'Keluhan tidak valid.',
-                    ]
-                ],
-                'dokter' => [
-                    'rules' => 'required|string|max_length[100]',
-                    'errors' => [
-                        'required' => 'Nama dokter wajib diisi.',
-                        'string'  => 'Nama dokter tidak valid.',
-                    ]
-                ],
-            ];
+                ];
+            }
 
             if (! $this->validate($rules)) {
                 return view('kunjungan/form', [
                     'asuransi_pasien' => $asuransiPasien,
                     'asuransi' => $asuransi,
                     'pasien' => $pasien,
+                    'unit' => $unit,
+                    'empOnUnit' => $empOnUnit,
                     'validation' => $this->validator,
                     'oldInput' => $this->request->getPost()
                 ]);
             }
 
+            $asuransiId = null;
+
+            if ($metode === 'asuransi') {
+                $asuransiId = $this->request->getPost('asuransi_pasien_id');
+            }
+
             $this->kunjunganModel->save([
                 'pasien_id'             => $this->request->getPost('pasien_id'),
-                'asuransi_pasien_id'    => $this->request->getPost('asuransi_pasien_id'),
+                'asuransi_pasien_id'    => $asuransiId,
+                'unit_id'               => $this->request->getPost('unit_id'),
+                'dpjp'                  => $this->request->getPost('emp_on_unit_id'),
                 'tanggal_kunjungan'       => $this->request->getPost('tanggal_kunjungan'),
                 'metode_pembayaran'       => $this->request->getPost('metode_pembayaran'),
                 'keluhan'               => $this->request->getPost('keluhan'),
-                'dokter'                => $this->request->getPost('dokter'),
             ]);
 
             return redirect()->to('/kunjungan')->with('success', 'Data kunjungan pasien berhasil ditambahkan.');
@@ -131,6 +170,9 @@ class Kunjungan extends BaseController
             'asuransi_pasien' => $asuransiPasien,
             'asuransi' => $asuransi,
             'pasien' => $pasien,
+            'unit' => $unit,
+            'empOnUnit' => $empOnUnit,
+            'kunjungan' => null,
             'validation' => $this->validator,
         ]);
     }
@@ -141,21 +183,24 @@ class Kunjungan extends BaseController
         $pasien = $this->pasienModel->select('id,nik,nama')->findAll();
         $asuransi = $this->asuransiModel->select('id,nama_asuransi')->findAll();
         $kunjungan = $this->kunjunganModel->getWithRelations()->find($id);
+        $unit = $this->unitModel->select()->findAll();
+        $empOnUnit = $this->empOnUnitModel->getWithEmployee()->findAll();
+
         $asuransiPasien = $this->asuransiPasienModel
             ->select('asuransi_pasien.id, asuransi.nama_asuransi')
             ->join('asuransi', 'asuransi.id = asuransi_pasien.asuransi_id')
             ->where('asuransi_pasien.pasien_id', $id)
             ->findAll();
 
-        $data = [
+        return view('kunjungan/form', [
             'asuransi_pasien' => $asuransiPasien,
             'kunjungan' => $kunjungan,
             'asuransi' => $asuransi,
             'pasien' => $pasien,
+            'unit' => $unit,
+            'empOnUnit' => $empOnUnit,
             'validation' => $this->validation
-        ];
-
-        return view('kunjungan/form', $data);
+        ]);
     }
 
     public function update($id)
@@ -163,70 +208,46 @@ class Kunjungan extends BaseController
         $pasien = $this->pasienModel->select('id,nik,nama')->findAll();
         $asuransi = $this->asuransiModel->select('id,nama_asuransi')->findAll();
         $kunjungan = $this->kunjunganModel->find($id);
+        $unit = $this->unitModel->select()->findAll();
+        $rules = $this->rules;
 
-        $rules = [
-            'pasien_id' => [
-                'rules' => 'required|integer',
-                'errors' => [
-                    'required' => 'Pasien wajib dipilih.',
-                    'integer'  => 'ID pasien tidak valid.',
-                ]
-            ],
-            'asuransi_pasien_id' => [
+        $metode = $this->request->getPost('metode_pembayaran');
+
+        if ($metode === 'asuransi') {
+            $rules['asuransi_pasien_id'] = [
                 'rules' => 'required|integer',
                 'errors' => [
                     'required' => 'Asuransi wajib dipilih.',
                     'integer'  => 'ID asuransi tidak valid.',
                 ]
-            ],
-
-            'tanggal_kunjungan' => [
-                'rules' => 'required|valid_date',
-                'errors' => [
-                    'required' => 'Tanggal kunjungan wajib diisi.',
-                    'valid_date'  => 'Tanggal kunjungan tidak valid.',
-                ]
-            ],
-            'metode_pembayaran' => [
-                'rules' => 'required|string',
-                'errors' => [
-                    'required' => 'Metode pembayaran wajib diisi.',
-                    'string'  => 'Metode pembayaran tidak valid.',
-                ]
-            ],
-            'keluhan' => [
-                'rules' => 'required|string|max_length[255]',
-                'errors' => [
-                    'required' => 'Keluhan wajib diisi.',
-                    'string'  => 'Keluhan tidak valid.',
-                ]
-            ],
-            'dokter' => [
-                'rules' => 'required|string|max_length[100]',
-                'errors' => [
-                    'required' => 'Nama dokter wajib diisi.',
-                    'string'  => 'Nama dokter tidak valid.',
-                ]
-            ],
-        ];
+            ];
+        }
 
         if (! $this->validate($rules)) {
             return view('kunjungan/form', [
                 'kunjungan' => $kunjungan,
                 'asuransi' => $asuransi,
                 'pasien' => $pasien,
+                'unit' => $unit,
                 'validation' => $this->validator,
                 'oldInput' => $this->request->getPost()
             ]);
         }
 
+        $asuransiId = null;
+
+        if ($metode === 'asuransi') {
+            $asuransiId = $this->request->getPost('asuransi_pasien_id');
+        }
+
         $this->kunjunganModel->update($id, [
             'pasien_id'             => $this->request->getPost('pasien_id'),
-            'asuransi_pasien_id'    => $this->request->getPost('asuransi_pasien_id'),
+            'asuransi_pasien_id'    => $asuransiId,
+            'unit_id'               => $this->request->getPost('unit_id'),
+            'dpjp'                  => $this->request->getPost('emp_on_unit_id'),
             'tanggal_kunjungan'       => $this->request->getPost('tanggal_kunjungan'),
             'metode_pembayaran'       => $this->request->getPost('metode_pembayaran'),
             'keluhan'               => $this->request->getPost('keluhan'),
-            'dokter'                => $this->request->getPost('dokter'),
         ]);
 
         return redirect()->to('/kunjungan')->with('success', 'Data kunjungan pasien berhasil diperbarui.');
@@ -234,6 +255,12 @@ class Kunjungan extends BaseController
 
     public function delete($id)
     {
+        $data = $this->kunjunganModel->find($id);
+
+        if (!$data) {
+            throw new PageNotFoundException();
+        }
+
         $this->kunjunganModel->delete($id);
         return redirect()->to('/kunjungan')->with('success', 'Data kunjungan pasien berhasil dihapus.');
     }
